Use int, as 'long' is difficult to marshal, and frame buffer sizes will always fit in <2GB.

+++ Source/ARX/ARController.cpp
@@ -836,10 +836,10 @@ done:
 }
 
 int ARController::videoPush(int videoSourceIndex,
-                            ARUint8 *buf0p, long buf0Size, int buf0PixelStride, int buf0RowStride,
-                            ARUint8 *buf1p, long buf1Size, int buf1PixelStride, int buf1RowStride,
-                            ARUint8 *buf2p, long buf2Size, int buf2PixelStride, int buf2RowStride,
-                            ARUint8 *buf3p, long buf3Size, int buf3PixelStride, int buf3RowStride)
+                            ARUint8 *buf0p, int buf0Size, int buf0PixelStride, int buf0RowStride,
+                            ARUint8 *buf1p, int buf1Size, int buf1PixelStride, int buf1RowStride,
+                            ARUint8 *buf2p, int buf2Size, int buf2PixelStride, int buf2RowStride,
+                            ARUint8 *buf3p, int buf3Size, int buf3PixelStride, int buf3RowStride)
 {
     if (videoSourceIndex < 0 || videoSourceIndex > (m_videoSourceIsStereo ? 1 : 0)) return -1;
