+++ Source/ARX/ARVideo/External/videoExternal.cpp
@@ -768,10 +768,10 @@ done:
 }
 
 int ar2VideoPushExternal(AR2VideoParamExternalT *vid,
-                         ARUint8 *buf0p, long buf0Size, int buf0PixelStride, int buf0RowStride,
-                         ARUint8 *buf1p, long buf1Size, int buf1PixelStride, int buf1RowStride,
-                         ARUint8 *buf2p, long buf2Size, int buf2PixelStride, int buf2RowStride,
-                         ARUint8 *buf3p, long buf3Size, int buf3PixelStride, int buf3RowStride)
+                         ARUint8 *buf0p, int buf0Size, int buf0PixelStride, int buf0RowStride,
+                         ARUint8 *buf1p, int buf1Size, int buf1PixelStride, int buf1RowStride,
+                         ARUint8 *buf2p, int buf2Size, int buf2PixelStride, int buf2RowStride,
+                         ARUint8 *buf3p, int buf3Size, int buf3PixelStride, int buf3RowStride)
 {
     int ret = -1;
     if (!vid) return -1; // Sanity check.
