+++ Source/ARX/ARVideoSource.cpp
@@ -444,10 +444,10 @@ int ARVideoSource::videoPushInit(int width, int height, const char *pixelFormat,
     return (ar2VideoPushInit(m_vid, width, height, pixelFormat, cameraIndex, cameraPosition));
 }
 
-int ARVideoSource::videoPush(ARUint8 *buf0p, long buf0Size, int buf0PixelStride, int buf0RowStride,
-                               ARUint8 *buf1p, long buf1Size, int buf1PixelStride, int buf1RowStride,
-                               ARUint8 *buf2p, long buf2Size, int buf2PixelStride, int buf2RowStride,
-                               ARUint8 *buf3p, long buf3Size, int buf3PixelStride, int buf3RowStride)
+int ARVideoSource::videoPush(ARUint8 *buf0p, int buf0Size, int buf0PixelStride, int buf0RowStride,
+                             ARUint8 *buf1p, int buf1Size, int buf1PixelStride, int buf1RowStride,
+                             ARUint8 *buf2p, int buf2Size, int buf2PixelStride, int buf2RowStride,
+                             ARUint8 *buf3p, int buf3Size, int buf3PixelStride, int buf3RowStride)
 {
     if (deviceState != DEVICE_RUNNING && deviceState != DEVICE_GETTING_READY) return 0;
