+++ Source/ARX/ARX_c.cpp
@@ -300,10 +300,10 @@ int arwVideoPushInit(int videoSourceIndex, int width, int height, const char *pi
 }
 
 int arwVideoPush(int videoSourceIndex,
-                uint8_t *buf0p, long buf0Size, int buf0PixelStride, int buf0RowStride,
-                uint8_t *buf1p, long buf1Size, int buf1PixelStride, int buf1RowStride,
-                uint8_t *buf2p, long buf2Size, int buf2PixelStride, int buf2RowStride,
-                uint8_t *buf3p, long buf3Size, int buf3PixelStride, int buf3RowStride)
+                uint8_t *buf0p, int buf0Size, int buf0PixelStride, int buf0RowStride,
+                uint8_t *buf1p, int buf1Size, int buf1PixelStride, int buf1RowStride,
+                uint8_t *buf2p, int buf2Size, int buf2PixelStride, int buf2RowStride,
+                uint8_t *buf3p, int buf3Size, int buf3PixelStride, int buf3RowStride)
 {
     if (!gARTK) return -1;
 
@@ -1372,22 +1372,22 @@ JNIEXPORT jint JNICALL JNIFUNCTION(arwVideoPush(JNIEnv *env, jobject obj, jint v
         return -1;
     }
     uint8_t *buf0p = NULL, *buf1p = NULL, *buf2p = NULL, *buf3p = NULL;
-    long buf0Size = 0, buf1Size = 0, buf2Size = 0, buf3Size = 0;
+    int buf0Size = 0, buf1Size = 0, buf2Size = 0, buf3Size = 0;
     if (buf0) {
         buf0p = (uint8_t *)env->GetDirectBufferAddress(buf0);
-        buf0Size = env->GetDirectBufferCapacity(buf0);
+        buf0Size = (int)env->GetDirectBufferCapacity(buf0);
     }
     if (buf1) {
         buf1p = (uint8_t *)env->GetDirectBufferAddress(buf1);
-        buf1Size = env->GetDirectBufferCapacity(buf1);
+        buf1Size = (int)env->GetDirectBufferCapacity(buf1);
     }
     if (buf2) {
         buf2p = (uint8_t *)env->GetDirectBufferAddress(buf2);
-        buf2Size = env->GetDirectBufferCapacity(buf2);
+        buf2Size = (int)env->GetDirectBufferCapacity(buf2);
     }
     if (buf3) {
         buf3p = (uint8_t *)env->GetDirectBufferAddress(buf3);
-        buf3Size = env->GetDirectBufferCapacity(buf3);
+        buf3Size = (int)env->GetDirectBufferCapacity(buf3);
     }
 
     return gARTK->videoPush(videoSourceIndex, buf0p, buf0Size, buf0PixelStride, buf0RowStride, buf1p, buf1Size, buf1PixelStride, buf1RowStride, buf2p, buf2Size, buf2PixelStride, buf2RowStride, buf3p, buf3Size, buf3PixelStride, buf3RowStride);
