+++ Source/ARX/include/ARX/ARController.h
@@ -192,10 +192,10 @@ public:
 
     int videoPushInit(int videoSourceIndex, int width, int height, const char *pixelFormat, int cameraIndex, int cameraPosition);
     int videoPush(int videoSourceIndex,
-                  ARUint8 *buf0p, long buf0Size, int buf0PixelStride, int buf0RowStride,
-                  ARUint8 *buf1p, long buf1Size, int buf1PixelStride, int buf1RowStride,
-                  ARUint8 *buf2p, long buf2Size, int buf2PixelStride, int buf2RowStride,
-                  ARUint8 *buf3p, long buf3Size, int buf3PixelStride, int buf3RowStride);
+                  ARUint8 *buf0p, int buf0Size, int buf0PixelStride, int buf0RowStride,
+                  ARUint8 *buf1p, int buf1Size, int buf1PixelStride, int buf1RowStride,
+                  ARUint8 *buf2p, int buf2Size, int buf2PixelStride, int buf2RowStride,
+                  ARUint8 *buf3p, int buf3Size, int buf3PixelStride, int buf3RowStride);
     int videoPushFinal(int videoSourceIndex);
 
 	/**
